Copy global verilog and cpp sources to local optional

+++ proj/clean_up.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
-rm -rf $1/src
-rm -rf $1/verilog_src
-rm -rf $1/cfu.v
+if [ $CLEAN_LOCAL_SOURCES = "1" ]; then
+    rm -rf $1/src
+    rm -rf $1/verilog_src
+    rm -rf $1/cfu.v
+fi
